+++ docs/Changelog.md
@@ -2,5 +2,11 @@
 
 All notable changes to this project will be documented in this file. This project uses [semantic versioning](https://semver.org/).
 
+## 0.2.0 (2024-03-06)
+  * New feature: unlock support.
+  * New feature: better device enumeration.
+  * New feature: access to the realtime events API for UniFi Access.
+  * Housekeeping.
+
 ## 0.1.0 (2024-03-02)
   * Initial release. Support for device enumeration and some foundational capabilities.